Making everything work with MSVC
- started to switch to hpx::parallel::datapar algorithms (requires HPX Master)
- addressed some of the warnings
- adding .gitignore

+++ src/geometry.hpp
@@ -180,7 +180,7 @@ public:
 		return (i >> dim) & 1;
 	}
 	octant neighbor(const direction& dir) {
-		integer ci = ZERO;
+		integer ci = static_cast<integer>(ZERO);
 		for (auto& d : dimension::full_set()) {
 			const integer bit = integer(1) << integer(d);
 			if (dir[d] == 0) {
